requirements.txt
pandas
numpy
matplotlib
mplfinance
tqdm
pyarrow
pillow
